refactor(platform/core): relocate InteractionHistory port to @freeside/core — close platform->sietch reach-in (arrakis-shqm)
Coherence F4: InteractionRecord + InteractionHistoryProvider were declared in themes/sietch and imported by the platform adapter via a 4-level relative path, inverting the belt direction (platform depending on a theme). Moved both (byte-identical) to @freeside/core/ports; capability-mesh.ts imports+re-exports them; audit-query-service.ts:14 repoints to @freeside/core/ports. Result: zero platform->sietch import statements remain; belt direction is platform-core <- theme. Pure type relocation, no runtime change. Inline-verified; full /review+/audit recommended pre-merge.
